LayoutTests/fast/css/lines-clamp-parsing-expected.txt

Lines Clamp Parsing Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testParsing("linesClampTest", "2 1", "-webkit-lines-clamp") is "2 1"
PASS testParsing("linesClampTest", "1 1", "-webkit-lines-clamp") is "1 1"
PASS testParsing("linesClampTest", "20% 4", "-webkit-lines-clamp") is "20% 4"
PASS testParsing("linesClampTest", "3 30%", "-webkit-lines-clamp") is "3 30%"
PASS testParsing("linesClampTest", "50% 60%", "-webkit-lines-clamp") is "50% 60%"
PASS testParsing("linesClampTest", "1 3 foo", "-webkit-lines-clamp") is "1 3 foo"
PASS testParsing("linesClampTest", "inherit", "-webkit-lines-clamp") is "inherit"
PASS testParsing("linesClampTest", "1 a", "-webkit-lines-clamp") is "1 a"
PASS testParsing("linesClampTest", "none", "-webkit-lines-clamp") is "none"
PASS testParsing("linesClampTest", "foo", "-webkit-lines-clamp") is "none"
PASS testParsing("linesClampTest", "1 2 3", "-webkit-lines-clamp") is "none"
PASS testParsing("linesClampTest", "a 1", "-webkit-lines-clamp") is "none"
PASS testParsing("linesClampTest", "abc123", "-webkit-lines-clamp") is "none"
PASS testParsing("linesClampTest", "-1", "-webkit-lines-clamp") is "none"
PASS testParsing("linesClampTest", "50px", "-webkit-lines-clamp") is "none"
PASS successfullyParsed is true

TEST COMPLETE
